Update rabies_sq.md
Pretty much just minor suggestions. Looks good. I do think you could cut some of the background out (ie, reduce the first three paragraphs into one as a brief preamble to the work you're going to talk about) - but it's not necessary. Suggested reordering the description of cycling a bit; also wasn't sure what was added by the parenthetical use of the 'vigilance' phrasing so suggested cutting it. Hard to give comments in this format! Hope this works...